# algamoney-ui/openapi.json
{"openapi":"3.0.1","info":{"title":"Algamoney API","description":"API para gerenciamento financeiro feedbacks","contact":{"name":"Jameson Henrique","url":"https://www.linkedin.com/in/jamesonhenrique/","email":"jamesonhenrique14@gmail.com"},"version":"1.0"},"servers":[{"url":"http://localhost:8080/","description":"Local ENV"},{"url":"...","description":"PROD ENV"}],"security":[{"bearerAuth":[]}],"paths":{"/pessoas/{codigo}":{"get":{"tags":["pessoa-resource"],"operationId":"buscarPeloCodigo","parameters":[{"name":"codigo","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PessoaResponseDTO"}}}}}},"put":{"tags":["pessoa-resource"],"operationId":"atualizar","parameters":[{"name":"codigo","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PessoaRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PessoaResponseDTO"}}}}}},"delete":{"tags":["pessoa-resource"],"operationId":"remover","parameters":[{"name":"codigo","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/pessoas/{codigo}/ativo":{"put":{"tags":["pessoa-resource"],"operationId":"atualizarPropriedadeAtivo","parameters":[{"name":"codigo","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"boolean"}}},"required":true},"responses":{"204":{"description":"No Content"}}}},"/lancamentos/{codigo}":{"get":{"tags":["lancamento-resource"],"operationId":"buscarPeloCodigo_1","parameters":[{"name":"codigo","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LancamentoResponseDTO"}}}}}},"put":{"tags":["lancamento-resource"],"operationId":"atualizar_1","parameters":[{"name":"codigo","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LancamentoRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LancamentoResponseDTO"}}}}}},"delete":{"tags":["lancamento-resource"],"operationId":"remover_1","parameters":[{"name":"codigo","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/pessoas":{"get":{"tags":["pessoa-resource"],"operationId":"pesquisar","parameters":[{"name":"page","in":"query","required":false,"schema":{"minimum":0,"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"maximum":100,"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponsePessoaResponseDTO"}}}}}},"post":{"tags":["pessoa-resource"],"operationId":"criar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PessoaRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PessoaResponseDTO"}}}}}}},"/lancamentos":{"get":{"tags":["lancamento-resource"],"operationId":"filtrar","parameters":[{"name":"descricao","in":"query","required":false,"schema":{"type":"string"}},{"name":"dataVencimentoDe","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"dataVencimentoAte","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"page","in":"query","required":false,"schema":{"minimum":0,"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"maximum":100,"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponseLancamentoResponseDTO"}}}}}},"post":{"tags":["lancamento-resource"],"operationId":"criar_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LancamentoRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LancamentoResponseDTO"}}}}}}},"/categorias":{"get":{"tags":["categoria-resource"],"operationId":"listar","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Categoria"}}}}}}},"post":{"tags":["categoria-resource"],"operationId":"criar_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Categoria"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Categoria"}}}}}}},"/auth/register":{"post":{"tags":["authentication-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/auth/authentication":{"post":{"tags":["authentication-controller"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/lancamentos/resumir":{"get":{"tags":["lancamento-resource"],"operationId":"resumir","parameters":[{"name":"descricao","in":"query","required":false,"schema":{"type":"string"}},{"name":"dataVencimentoDe","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"dataVencimentoAte","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResumoLancamento"}}}}}}},"/categorias/{codigo}":{"get":{"tags":["categoria-resource"],"operationId":"buscarPeloCodigo_2","parameters":[{"name":"codigo","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Categoria"}}}}}}}},"components":{"schemas":{"Endereco":{"type":"object"},"PessoaRequestDTO":{"required":["ativo","endereco","nome"],"type":"object","properties":{"nome":{"type":"string"},"endereco":{"$ref":"#/components/schemas/Endereco"},"ativo":{"type":"boolean"}}},"PessoaResponseDTO":{"type":"object","properties":{"codigo":{"type":"integer","format":"int64"},"nome":{"type":"string"},"endereco":{"$ref":"#/components/schemas/Endereco"},"ativo":{"type":"boolean"}}},"Categoria":{"required":["nome"],"type":"object","properties":{"codigo":{"type":"integer","format":"int64"},"nome":{"maxLength":20,"minLength":3,"type":"string"}}},"LancamentoRequestDTO":{"required":["categoria","dataVencimento","descricao","pessoa","tipo","valor"],"type":"object","properties":{"descricao":{"type":"string"},"dataVencimento":{"type":"string","format":"date"},"dataPagamento":{"type":"string","format":"date"},"valor":{"type":"number"},"observacao":{"type":"string"},"tipo":{"type":"string","enum":["RECEITA","DESPESA"]},"categoria":{"$ref":"#/components/schemas/Categoria"},"pessoa":{"$ref":"#/components/schemas/Pessoa"}}},"Pessoa":{"required":["ativo","endereco","nome"],"type":"object","properties":{"codigo":{"type":"integer","format":"int64"},"nome":{"type":"string"},"endereco":{"$ref":"#/components/schemas/Endereco"},"ativo":{"type":"boolean"}}},"LancamentoResponseDTO":{"type":"object","properties":{"codigo":{"type":"integer","format":"int64"},"descricao":{"type":"string"},"dataVencimento":{"type":"string","format":"date"},"dataPagamento":{"type":"string","format":"date"},"valor":{"type":"number"},"observacao":{"type":"string"},"tipo":{"type":"string","enum":["RECEITA","DESPESA"]},"categoria":{"$ref":"#/components/schemas/Categoria"},"pessoa":{"$ref":"#/components/schemas/Pessoa"}}},"RegisterRequest":{"type":"object","properties":{"nome":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"AuthenticationResponse":{"type":"object","properties":{"token":{"type":"string"}}},"AuthenticationRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"PageResponsePessoaResponseDTO":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/PessoaResponseDTO"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}},"PageResponseLancamentoResponseDTO":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/LancamentoResponseDTO"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"PageResumoLancamento":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/ResumoLancamento"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageSize":{"type":"integer","format":"int32"},"pageNumber":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"}}},"ResumoLancamento":{"type":"object"},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT authentication description","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}